refactor: presentation mixin の引数を減らした
